Project4 : Django RESTful V1.0

+++ Project2/crud/accounts/views.py
@@ -1,19 +1,18 @@
-# accounts/views.py
-
 from django.shortcuts import render, redirect
 from django.contrib.auth import login as auth_login
 from django.contrib.auth import logout as auth_logout
-from django.contrib.auth.forms import AuthenticationForm, UserCreationForm, PasswordChangeForm
-from .forms import CustomUserChangeForm, CustomUserCreationForm
 from django.contrib.auth import update_session_auth_hash
-from django.views.decorators.http import require_POST, require_http_methods
 from django.contrib.auth.decorators import login_required
+from django.views.decorators.http import require_POST, require_http_methods
+from django.contrib.auth.forms import AuthenticationForm, PasswordChangeForm
+from .forms import CustomUserCreationForm, CustomUserChangeForm
 
-
+# Create your views here.
 @require_http_methods(['GET', 'POST'])
 def login(request):
     if request.user.is_authenticated:
-        return redirect('articles/index')
+        return redirect('articles:index')
+
     if request.method == 'POST':
         form = AuthenticationForm(request, request.POST)
         # form = AuthenticationForm(request, data=request.POST)
@@ -28,50 +27,65 @@ def login(request):
     }
     return render(request, 'accounts/login.html', context)
 
+
 @require_POST
 def logout(request):
-    auth_logout(request)
+    if request.user.is_authenticated:
+        auth_logout(request)
     return redirect('articles:index')
+
 
 @require_http_methods(['GET', 'POST'])
 def signup(request):
-    if request.method == "POST":
+    if request.user.is_authenticated:
+        return redirect('articles:index')
+        
+    if request.method == 'POST':
         form = CustomUserCreationForm(request.POST)
         if form.is_valid():
             user = form.save()
+            # 회원가입 후 로그인
             auth_login(request, user)
             return redirect('articles:index')
     else:
         form = CustomUserCreationForm()
     context = {
-        'form' : form
+        'form': form,
     }
     return render(request, 'accounts/signup.html', context)
 
+
+@require_POST
 def delete(request):
-    request.user.delete()
-    auth_logout(request)
+    if request.user.is_authenticated:
+        request.user.delete()
+        auth_logout(request)
     return redirect('articles:index')
+
 
 @login_required
 @require_http_methods(['GET', 'POST'])
 def update(request):
-    if request.method == "POST":
+    if request.method == 'POST':
         form = CustomUserChangeForm(request.POST, instance=request.user)
+        # form = CustomUserChangeForm(data=request.POST, instance=request.user)
         if form.is_valid():
             form.save()
             return redirect('articles:index')
     else:
         form = CustomUserChangeForm(instance=request.user)
     context = {
-        'form' : form,
+        'form': form,
     }
     return render(request, 'accounts/update.html', context)
 
+
+@login_required
 @require_http_methods(['GET', 'POST'])
 def change_password(request):
     if request.method == 'POST':
         form = PasswordChangeForm(request.user, request.POST)
+        # form = PasswordChangeForm(user=request.user, data=request.POST)
         if form.is_valid():
             form.save()
             update_session_auth_hash(request, form.user)
@@ -79,6 +93,6 @@ def change_password(request):
     else:
         form = PasswordChangeForm(request.user)
     context = {
-        'form' : form,
+        'form': form,
     }
     return render(request, 'accounts/change_password.html', context)
